Refactor to abstract out sync specs versus common spec functionality
