+++ src/Polly.Caching.Distributed.SharedSpecs/Integration/CacheRoundTripSpecsBase.cs
@@ -7,6 +7,50 @@ namespace Polly.Caching.Distributed.Specs.Integration
     {
         protected const string OperationKey = "SomeOperationKey";
 
+        public abstract void Should_roundtrip_this_variant_of<TResult>(TResult testValue);
+
+        [Theory]
+        [MemberData(nameof(SampleClassData))]
+        public void Should_roundtrip_all_variants_of_reference_type(SampleClass testValue)
+        {
+            Should_roundtrip_this_variant_of<SampleClass>(testValue);
+        }
+
+        [Theory]
+        [MemberData(nameof(SampleStringData))]
+        public void Should_roundtrip_all_variants_of_string(String testValue)
+        {
+            Should_roundtrip_this_variant_of<String>(testValue);
+        }
+
+        [Theory]
+        [MemberData(nameof(SampleNumericData))]
+        public void Should_roundtrip_all_variants_of_numeric(int testValue)
+        {
+            Should_roundtrip_this_variant_of<int>(testValue);
+        }
+
+        [Theory]
+        [MemberData(nameof(SampleEnumData))]
+        public void Should_roundtrip_all_variants_of_enum(SampleEnum testValue)
+        {
+            Should_roundtrip_this_variant_of<SampleEnum>(testValue);
+        }
+
+        [Theory]
+        [MemberData(nameof(SampleBoolData))]
+        public void Should_roundtrip_all_variants_of_bool(bool testValue)
+        {
+            Should_roundtrip_this_variant_of<bool>(testValue);
+        }
+
+        [Theory]
+        [MemberData(nameof(SampleNullableBoolData))]
+        public void Should_roundtrip_all_variants_of_nullable_bool(bool? testValue)
+        {
+            Should_roundtrip_this_variant_of<bool?>(testValue);
+        }
+
         public static TheoryData<SampleClass> SampleClassData =>
             new TheoryData<SampleClass>
             {
