+++ src/Polly.Caching.Distributed.SharedSpecs/Integration/CacheRoundTripSpecs_NetStandardIDistributedCacheProvider_Sync_ByteArray.cs
@@ -1,0 +1,8 @@
+﻿#if !NETCOREAPP1_1
+
+namespace Polly.Caching.Distributed.Specs.Integration
+{
+    public class CacheRoundTripSpecs_NetStandardIDistributedCacheProvider_Sync_ByteArray : CacheRoundTripSpecsSyncBase<byte[]> { }
+}
+
+#endif
